fixed minor style things, added 100 vh, moved details code into pageDetails
