+++ src/styles/components/_about.scss
@@ -14,7 +14,7 @@
     margin: 0 auto;
     flex-direction: column;
     text-align: center;
-    height: 75vh;
+    height: 100vh;
     max-width: 40rem;
     padding: 1.5rem;
