+++ src/styles/components/_card.scss
@@ -151,7 +151,7 @@
                 flex-direction: column;
                 align-items: center;
                 justify-content: center;
-                opacity: 1;
+                opacity: 0;
                 transition: $transiting; 
                 text-align: center;
                 border-radius: 1.25rem;
