+++ src/styles/components/_favourites.scss
@@ -8,7 +8,7 @@
     justify-content: center;
     align-items: center;
     text-align: center;
-    height: 75vh;
+    height: 100vh;
     padding: 1rem;
     background-color: $black-color;
     text-transform: uppercase;
@@ -21,7 +21,7 @@
 
 .favourites {
     padding: 0.6rem;
-    min-height: 75vh;
+    min-height: 100vh;
 }
 
 
@@ -37,7 +37,7 @@
     }
     .favourites {
         padding: 2rem;
-        min-height: 75vh;
+        min-height: 100vh;
     }
 }
